rename backend/corpora to backend/portal
