+++ Dockerfile.handle_dataset_submissions
@@ -8,7 +8,6 @@ ENV PIP_NO_CACHE_DIR=1
 RUN pip3 install -r requirements.txt
 
 COPY backend/__init__.py ./backend/__init__.py
-COPY backend/corpora/__init__.py ./backend/corpora/__init__.py
 COPY backend/common ./backend/common
 
 ARG HAPPY_BRANCH="unknown"
