+++ Dockerfile.handle_success
@@ -5,7 +5,6 @@ COPY backend/upload_success .
 RUN pip3 install -r requirements.txt
 
 COPY backend/__init__.py ./backend/__init__.py
-COPY backend/corpora/__init__.py ./backend/corpora/__init__.py
 COPY backend/common ./backend/common
 COPY backend/dataset_processing ./backend/dataset_processing
