+++ backend/portal/api/v1/curation/collections/collection_id/datasets/status.py
@@ -1,7 +1,7 @@
 from flask import g, make_response, jsonify
 
 from backend.api_server.db import dbconnect
-from backend.corpora.api.v1.common import get_dataset_else_error
+from backend.portal.api.v1.common import get_dataset_else_error
 
 
 @dbconnect
